requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
email-validator==2.0.0.post2
Pillow==10.0.0
qrcode==7.4.2
python-dotenv==1.0.0
requests==2.31.0

# Для отладки
Flask-DebugToolbar==0.13.1
